requirements.txt
Django>=4.2.0
django-crontab>=0.7.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
dj-database-url>=1.0.0
django-allauth>=0.57.0
celery==5.5.3
weasyprint==60.1
reportlab>=4.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0    